refactor: 메서드 네이밍 명확화 및 패키지 구조 개선
