+++ src/main/java/com/techfork/global/common/code/BaseCode.java
@@ -1,0 +1,7 @@
+package com.techfork.global.common.code;
+
+import com.techfork.global.response.ReasonDTO;
+
+public interface BaseCode {
+    ReasonDTO getReason();
+}
