--- src/main/java/com/techfork/global/response/code/BaseCode.java
@@ -1,7 +1,0 @@
-package com.techfork.global.response.code;
-
-import com.techfork.global.response.dto.ReasonDTO;
-
-public interface BaseCode {
-    ReasonDTO getReason();
-}
